Messaging System complete

+++ components/Message.jsx
@@ -7,6 +7,8 @@ const Message = ({ message }) => {
   const [isRead, setIsRead] = useState(message.read);
   const [isDeleted, setIsDeleted] = useState(false);
 
+  const { unreadCount, setUnreadCount } = useGlobalContext();
+
   // const { setUnreadCount } = useGlobalContext();
 
   const handleReadClick = async () => {
@@ -18,7 +20,7 @@ const Message = ({ message }) => {
       if (res.status === 200) {
         const { read } = await res.json();
         setIsRead(read);
-        // setUnreadCount((prevCount) => (read ? prevCount - 1 : prevCount + 1));
+        setUnreadCount((prevCount) => (read ? prevCount - 1 : prevCount + 1));
         if (read) {
           toast.success('Marked as read');
         } else {
@@ -32,14 +34,20 @@ const Message = ({ message }) => {
   };
 
   const handleDeleteClick = async () => {
+    const confirmed = window.confirm(
+      'Are you sure that you want to delete this message?'
+    );
+
+    if (!confirmed) return;
+
     try {
       const res = await fetch(`/api/messages/${message._id}`, {
         method: 'DELETE',
       });
 
-      if (res.status === 200) {
+      if (res.status === 201) {
         setIsDeleted(true);
-        // setUnreadCount((prevCount) => prevCount - 1);
+        setUnreadCount((prevCount) => prevCount - 1);
         toast.success('Message Deleted');
       }
     } catch (error) {
@@ -72,7 +80,11 @@ const Message = ({ message }) => {
 
         <li>
           <strong>Reply Email:</strong>{' '}
-          <a href={`mailto:${message.email}`} className='text-blue-500'>
+          <a
+            href={`mailto:${message.email}`}
+            target='_blank'
+            className='text-blue-500'
+          >
             {message.email}
           </a>
         </li>
